Split code into separate components / files
