+++ src/pages/visit.tsx
@@ -1,4 +1,5 @@
 import Header from '@/components/Header'
+import IncidentForm from '@/components/IncidentForm'
 
 const Visit = () => {
   return (
@@ -8,6 +9,8 @@ const Visit = () => {
 
       <main>
         <p>Visit Page</p>
+        {/* TODO: Add button to open up form*/}
+        <IncidentForm isVetVisit={false} />
       </main>
     </>
   )
