+++ src/types.ts
@@ -1,0 +1,10 @@
+export type IncidentForm = {
+  userID: string
+  userName: string
+  email: string
+  petName: string
+  vetName?: string
+  date: Date
+  time: Date
+  details: string
+}
